refactor Converter to function, restructure folders
