+++ scripts/Classes/Recorder.js
