+++ scripts/Classes/Storage.js
